LPS-76972 Rename to Parser
