+++ modules/apps/web-experience/fragment/fragment-entry-processor/fragment-entry-processor-editable/src/main/java/com/liferay/fragment/entry/processor/editable/parser/impl/ImgEditableElementParser.java
@@ -12,23 +12,28 @@
  * details.
  */
 
-package com.liferay.fragment.entry.processor.editable.replacer;
+package com.liferay.fragment.entry.processor.editable.parser.impl;
 
+import com.liferay.fragment.entry.processor.editable.parser.EditableElementParser;
+import com.liferay.portal.kernel.xml.Attribute;
 import com.liferay.portal.kernel.xml.Element;
 
+import org.osgi.service.component.annotations.Component;
+
 /**
- * This service provides a utility to replace editable element value.
- *
  * @author Pavel Savinov
  */
-public interface EditableElementReplacer {
+@Component(
+	immediate = true, property = {"editable.tag.name=img"},
+	service = EditableElementParser.class
+)
+public class ImgEditableElementParser implements EditableElementParser {
 
-	/**
-	 * Replaces editable element value with the provided one.
-	 *
-	 * @param element Editable element to replace
-	 * @param value New element value
-	 */
-	public void replace(Element element, String value);
+	@Override
+	public void replace(Element element, String value) {
+		Attribute attribute = element.attribute("src");
+
+		attribute.setValue(value);
+	}
 
 }
